pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragas-gui"
version = "0.2.0"
description = "Streamlit-based evaluation UI for Ragas RAG pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "ragas>=0.4",
    "streamlit>=1.30",
    "pandas",
    "plotly",
    "openai",
    "datasets",
    "nest_asyncio",
    "langchain-openai",
    "langchain-community",
]

[project.optional-dependencies]
telemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-instrumentation-openai>=0.40",
]
dev = [
    "pytest>=7.0",
    "pytest-bdd>=7.0",
    "ruff",
]
all = [
    "ragas-gui[telemetry,dev]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
bdd_features_base_dir = "tests/features"
markers = [
    "bdd: BDD scenario tests",
    "unit: unit tests",
    "validation: data upload and validation tests",
    "metrics: metric configuration tests",
    "observability: telemetry and observability tests",
    "llm_config: LLM and embeddings configuration tests",
]
